Version 0.1.0 - Initial Git Commit
I'll start the Git history at version 0.1.0, which was the first version
uploaded to Factorio's Mod Portal. I'm doing this before I start working
on some new features.

Note that this repository doesn't (yet) contain the code used to create
the LUTs. That code is in a bit of a messy state, and basically when I
create a new LUT I go to one of my Shadertoy shaders and mess with it
until I have a LUT, and then I don't save changes.

I plan to change that, and part of that effort is visible via changes I
made to the directory structure, to separate files that get included
from files that process and generate those files.

For now, the actual LUTs are included under res/graphics. That might not
be the case after I have the code to generate the LUTs put in here in a
way that can be programmatically run to generate the files.